feat: add AuthorService interface
The interface AuthorService is added to provide a new level of abstraction
